System.Clock.Seconds

Documentation

data Clock

Clock types. A clock may be system-wide (that is, visible to all processes) or per-process (measuring time that is meaningful only within a process). All implementations shall support Realtime.

Constructors

Monotonic
The identifier for the system-wide monotonic clock, which is defined as a clock measuring real time, whose value cannot be set via clock_settime and which cannot have negative clock jumps. The maximum possible clock jump shall be implementation defined. For this clock, the value returned by getTime represents the amount of time (in seconds and nanoseconds) since an unspecified point in the past (for example, system start-up time, or the Epoch). This point does not change after system start-up time. Note that the absolute value of the monotonic clock is meaningless (because its origin is arbitrary), and thus there is no need to set it. Furthermore, realtime applications can rely on the fact that the value of this clock is never set. (Identical to Boottime since Linux 4.17, see https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d6ed449afdb38f89a7b38ec50e367559e1b8f71f) CLOCK_MONOTONIC (macOS - SYSTEM_CLOCK)

Realtime
The identifier of the system-wide clock measuring real time. For this clock, the value returned by getTime represents the amount of time (in seconds and nanoseconds) since the Epoch. CLOCK_REALTIME (macOS - CALENDAR_CLOCK, Windows - GetSystemTimeAsFileTime)

ProcessCPUTime
The identifier of the CPU-time clock associated with the calling process. For this clock, the value returned by getTime represents the amount of execution time of the current process.

ThreadCPUTime
The identifier of the CPU-time clock associated with the calling OS thread. For this clock, the value returned by getTime represents the amount of execution time of the current OS thread.

MonotonicRaw
(since Linux 2.6.28, macOS 10.12) Similar to Monotonic, but provides access to a raw hardware-based time that is not subject to NTP adjustments or the incremental adjustments performed by adjtime(3). CLOCK_MONOTONIC_RAW (Windows - QueryPerformanceCounter, QueryPerformanceFrequency)

Boottime
(since Linux 2.6.39; Linux-specific) Identical to Monotonic, except it also includes any time that the system is suspended. This allows applications to get a suspend-aware monotonic clock without having to deal with the complications of Realtime, which may have discontinuities if the time is changed using settimeofday(2). (since Linux 4.17; identical to Monotonic) CLOCK_BOOTTIME

MonotonicCoarse
(since Linux 2.6.32; Linux-specific) A faster but less precise version of Monotonic. Use when you need very fast, but not fine-grained timestamps. CLOCK_MONOTONIC_COARSE

RealtimeCoarse
(since Linux 2.6.32; Linux-specific) A faster but less precise version of Realtime. Use when you need very fast, but not fine-grained timestamps. CLOCK_REALTIME_COARSE

newtype Seconds

Constructors

Seconds

Fields

toTimeSpec :: TimeSpec

getTime :: Clock -> IO Seconds

The getTime function shall return the current value for the specified clock.

getRes :: Clock -> IO Seconds

The getRes function shall return the resolution of any clock. Clock resolutions are implementation-defined and cannot be set by a process.

fromNanoSecs :: Integer -> Seconds

Seconds from nano seconds.

toNanoSecs :: Seconds -> Integer

Seconds to nano seconds.

diffTimeSpec :: Seconds -> Seconds -> Seconds

Compute the absolute difference.